pyproject.toml
[project]
name = "chi2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "mcp[cli]>=1.0.0",
    "numpy>=2.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "umap-learn>=0.5.7",
]
